Moving the XML overlay merge into the serialization package.
- ContentXmlMerger becomes Rebellion.Util.Serialization.XmlOverlay.
- GameSerializer gains an XmlDocument overload that validates through the
  configured schema set, removing the save-and-reparse round-trip.
- LoadGameConfig now only resolves paths and layers the optional pack
  override over the application defaults.
